Se solucionan problemas con las relaciones de las entidades y se agrega el controlador de chromosome

+++ src/main/java/com/breaze/genomebank/genome/infraestructure/repository/IGenomeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IGenomeRepository extends JpaRepository<Genome, Long> {
+public interface IGenomeRepository extends JpaRepository<Genome, Integer> {
 }
